refactor: components to app
