+++ src/app/services/ComponentInfos.tsx
